Remove `success_message` from page template helper
Rubocop raised a self assignment issue in the `_uploaded_files` partials
regarding the `success_message` variable.

On investigation, it looks like the success message pattern is used in only one
place (the feedback page). Its use was unconventional, and didn't seem like
something the top level `page_template` helper should be concerned with either.

This replaces the current usage with a more conventional flash notice.

The behaviour is the same (i.e providers are redirected to the feedback page and
shown a message after signing out), but the style of the message has changed
slightly.

Following discussion with design, it was agreed that the [notification
component] would be used for all flash notice messages.

This also brings mock saml and non-mock saml behvaiour inline, so that both sets
of use cases mean the user is redirected to the feedback page.

[notification component]: https://design-system.service.gov.uk/components/notification-banner/