Fix flaky test
Before, the feedback page used the `.enum_radio_buttons` class method from the
`TranslatableModelAttribute` concern to build options for the difficulty and
satisfaction feedback questions.

While these provided a nice API for building the options, it added unnecessary
complexity and opened cache-sized door of opportunity for bugs.

This was highlighted by the [flaky test] which started failing intermittently
(but with very high frequency) since tests were broken down in #4852.

When the feedback form was completed in Welsh, the translations were cached and
subsequent tests of the feedback form in English would fail because the "Easy"
option was missing (it was "ysaE" instead).

This fixes the flaky test by replacing the use of `.enum_radio_buttons` with an
inline translation lookup —removing the cache and source of the leak.

[flaky test]: https://app.circleci.com/pipelines/github/ministryofjustice/laa-apply-for-legal-aid/20171/workflows/ba644a60-8077-42db-97fe-f1d733948eb1/jobs/104965