Update Ruby to v3.2.0 (#4893)
* Update Ruby to v3.2.0

Merry (belated) Christmas.

Ruby 3.2 has been [released], and this updates to the latest version.

[released]: https://www.ruby-lang.org/en/news/2022/12/25/ruby-3-2-0-released/

* Refactor `Proceedings::ClientInvolvementTypeForm`

Before, the chained use of `#filter_map` and `#reduce` parsed an array of
`ClientInvolvementTypeStruct` to retrieve the description for a given client
involvement type ccms code.

In Ruby 3.1.3, not passing any arguments or a block to `#inject` (i.e `#reduce`)
had the same effect as passing an empty block.

```ruby
[1] pry(#<Proceedings::ClientInvolvementTypeForm>)> client_involvement_types
=> [#<LegalFramework::ClientInvolvementTypes::All::ClientInvolvementTypeStruct:0x000000011a553500 @ccms_code="A", @description="Applicant/claimant/petitioner">,
 #<LegalFramework::ClientInvolvementTypes::All::ClientInvolvementTypeStruct:0x000000011a5534b0 @ccms_code="D", @description="Defendant/respondent">,
 #<LegalFramework::ClientInvolvementTypes::All::ClientInvolvementTypeStruct:0x000000011a553488 @ccms_code="W", @description="Subject of proceedings (child)">,
 #<LegalFramework::ClientInvolvementTypes::All::ClientInvolvementTypeStruct:0x000000011a553460 @ccms_code="I", @description="Intervenor">,
 #<LegalFramework::ClientInvolvementTypes::All::ClientInvolvementTypeStruct:0x000000011a553438 @ccms_code="Z", @description="Joined party">]
[2] pry(#<Proceedings::ClientInvolvementTypeForm>)> client_involvement_types.filter_map { |cit| cit.description if cit.ccms_code == client_involvement_type_ccms_code }
=> ["Applicant/claimant/petitioner"]
[3] pry(#<Proceedings::ClientInvolvementTypeForm>)> client_involvement_types.filter_map { |cit| cit.description if cit.ccms_code == client_involvement_type_ccms_code }.reduce
=> "Applicant/claimant/petitioner"
```

In Ruby 3.2.0 this raises an error:

```
ArgumentError: wrong number of arguments (given 0, expected 1..2)
```

If an empty block is passed then no error is raised.

```ruby
[1] pry(#<Proceedings::ClientInvolvementTypeForm>)> client_involvement_types
=> [#<LegalFramework::ClientInvolvementTypes::All::ClientInvolvementTypeStruct:0x000000011839c0d8 @ccms_code="A", @description="Applicant/claimant/petitioner">,
 #<LegalFramework::ClientInvolvementTypes::All::ClientInvolvementTypeStruct:0x000000011839c088 @ccms_code="D", @description="Defendant/respondent">,
 #<LegalFramework::ClientInvolvementTypes::All::ClientInvolvementTypeStruct:0x000000011839c060 @ccms_code="W", @description="Subject of proceedings (child)">,
 #<LegalFramework::ClientInvolvementTypes::All::ClientInvolvementTypeStruct:0x000000011839c038 @ccms_code="I", @description="Intervenor">,
 #<LegalFramework::ClientInvolvementTypes::All::ClientInvolvementTypeStruct:0x000000011839c010 @ccms_code="Z", @description="Joined party">]
[2] pry(#<Proceedings::ClientInvolvementTypeForm>)> client_involvement_types.filter_map { |cit| cit.description if cit.ccms_code == client_involvement_type_ccms_code }
=> ["Applicant/claimant/petitioner"]
[3] pry(#<Proceedings::ClientInvolvementTypeForm>)> client_involvement_types.filter_map { |cit| cit.description if cit.ccms_code == client_involvement_type_ccms_code }.reduce
ArgumentError: wrong number of arguments (given 0, expected 1..2)
from (pry):3:in `reduce'
[4] pry(#<Proceedings::ClientInvolvementTypeForm>)> client_involvement_types.filter_map { |cit| cit.description if cit.ccms_code == client_involvement_type_ccms_code }.reduce {}
=> "Applicant/claimant/petitioner"
```

The reason for this change in behvaiour was not immediately clear to me from the
release notes, but I suspect it might be something to do with the changes to
arguments handling in Ruby 3.2.0 —but I'm really not sure.

This refactors the form so that `#find` retrieves the matching client
involvement type, which improves readability and resolves the issue.

Since there should only ever be one matching client involvement type (and
beforehand we weren't handling multiple matches anyway), this seems like a
reasaonble approach.

A `before_validation` callback has been introduced in place of the overriden
save method. This felt like a nice approach (and one that we use elsewhere),
but it does introduce the risk of descriptions not being set if for some reason
we start saving data without validations.

The conditional comparison against `client_involvement_type_ccms_code` didn't do
what it intended, since a form will submit a blank (empty) string rather than
`nil`. Updating the conditional to check for presence resolves this.