Stop using Webpacker for styles and assets
Before, assets and styles were managed by webpacker.

This is a first step in migrating away from webpacker (which is deprecated) in
favour of one of the new Rails 7 defaults.

[Propshaft](https://github.com/rails/propshaft) is introduced as an asset
pipeline, while [cssbundling-rails](https://github.com/rails/cssbundling-rails)
uses [sass](https://github.com/sass/sass) to build our CSS.

Build tasks are attached to `assets:precompile` by default, and changes can be
watched in development.

Rather than keeping copies of `govuk-frontend` assets in the repo, the files in
the JS package are referenced. This was always possible, but for some reason we
weren't doing it.

Stylesheets and imports have also been tidied up and simplified, with duplicate
imports removed.