Replace webpacker with jsbundling-rails
This is the next step in migrating away from webpacker.

Since Rails 7, we can use
[jssbundling-rails](https://github.com/rails/jsbundling-rails) with a tool of
our choice to bundle our JavaScript and deliver it via the asset pipeline.

This replaces webpack with [esbuild](https://github.com/evanw/esbuild) which is
fast, simple, and easy to use.

A build task attaches itself to `assets:precompile` by default, and changes can
again be watched in development.

---

This necessitated some simple changes, and some things have also been tidied up:

- Removing webpack means we need to explicity depend on `core-js` for some
  polyfills.

- Removing webpack means we need to explicitly add some `babel` packages as
  development dependencies —since `jest` requires them.

- Since we are not using direct uploads, there is no need for us to install the
  `@rails/active-storage` package. This has been removed.

- We no longer need `webpack-dev-server` (it's been broken forever anyway), and
  this means we can remove localhost from the content security policy in
  development.

- Removing webpack means changes to the way we import our JS into
  `application.js`.

- Remove all webpack(er) binstubs and config.

- Co-locate JS test files with the JS files.

---

By default, the earliest `esbuild` can transpile to is ES6 JavaScript. Since
IE11 is officially end-of-life and not supported, we took the decision to drop
support for it (i.e ES6 is fine).

If we decide we want to bundle IE11 compatible JS, then we can add an `esbuild`
config file, and use `babel` to transpile compatible code. We might also need to
re-add `regenerator-runtime` for compatible async JS.

---

We are still reliant on `jquery` to tell us if there are any ongoing AJAX
requests in tests. Previous commits experimented with removing this dependency
with some custom `axios` interceptors, but a decision was made to revisit as
part of a wider JS-focused refactor.