Tweak and extend application DoB specs
Came of investigating bug AP-3782

Previous use of `3.days.from_now`, which returns a Timestamp, in
the test was actually raising and invalid date error. Actual behaviour
in the app for dates in the future is to raise date_is_in_future error
with message "Date of birth must be in the past". Replacing with
`Date.tomorrow` excercises the intended logic.

Extended tests to cover far in past dates too.