Migrate applicant SecureData true layer tokens
Before, the bespoke `SecureData` pattern was used to store applicant TrueLayer
tokens.

In #4967, a new `encrypted_true_layer_token` encrypted attribute was introduced,
and data was written to it as part of the TrueLayer flow.

This adds a temporary idempotent task to copy encrypted token data to the new
attribute, and updates existing methods to read from the new attribute if it
is present.

[Link to story](https://dsdmoj.atlassian.net/browse/AP-3743)