Explictly use redis instead of redis-client
This is because we are using redis-namespace which
there causes this error to be raised on the server during
deployment

```shell
Your Redis configuration uses the namespace 'bump-sidekiq-6-to-7-with-forke-applyforlegalaid-uat.cloud-platform.service.justice.gov.uk' but this feature isn't supported by redis-client. Either use the redis adapter or remove the namespace.
```