Add basic under 18 for means test purposes
This can be used to display alternate declaration
content for uunder 18s prior to MTR phase one go live
but, note, it does not use the same age calculation
as #age does. We are looking to harmonise all age calculation
to use age_for_means_test purposes in the foreseeeable future
however so this should be good to go, baring rare edge cases.