Ignore the `SecureData` true layer token
Before #4698, the `true_layer_secure_data` association was used to retrieve
secure TrueLayer tokens when interacting with the TrueLayer API to import
client's bank data.

Now that existing tokens have been migrated to the new
`encrypted_true_layer_token` column and the application is using this data, we
can begin to remove the legacy true layer token code.

This ignores the `true_layer_secure_data_id` column as a first step, and that
change was used to drive the work —fixing (and refactoring) failing tests as
they arose.

As part of the work, bank providers no longer store the TrueLayer `token` and
`token_expires_at` attributes. These were not used anywhere, and can be accessed
through the associated applicant if required in future.

When fixing the `ImportProviderService` specs, a uniqueness validation (and
corresponding index) was added to ensure duplicate bank providers cannot be
added for a given applicant. #1748 did the initial work to resolve a bug, but
this validation was omitted.

[Link to story](https://dsdmoj.atlassian.net/browse/AP-3743)