Amend lookup for postgres database and password
Changed in the bitnami/postgresql Chart for version ^11.0.0

The way Bitnami handles postgres inputs was changed in v11.
This moves the secrets into the new, .auth, structure and then
updates secret.yaml to read from the new location where appropriate
The username is outside of the .auth structure as postgres does not
expect a postgres username entry, so we will only use it to populate
the secrets for for the web instances.