Introduce encrypted Citizen access tokens
In #5010 we started encrypting a `citizen_url_id` and storing it alongside the
date the corresponding Citizen application link would expire on the
`LegalAidApplication` model.

As work to migrate from the `SecureData` approach continued in #5018, a bug was
discovered in manual testing.

Rather than generating a token for the Citizen URL once and passing that around,
tokens are generated multiple times for an application in the process of sending
an email to the client.

This means there could be X different `SecureData` records in the database for
an application, and we don't know which correspond to links sent in emails to
Citizens.

In order to accommodate this behaviour, and to migrate existing data so we can
continue to handle existing links appropriately -a new approach was needed.

This introduces the concept of a Citizen access token, that can be sent to
Citizens in emails, expire, and used to retrieve a client's legal aid
application.

This approach will also us to accommodate existing links (including expired
links), however we might later decide that initial requirements laid out in
the [initial ticket] and [discussion] or decisions taken in #73 are no longer
appropriate.

[initial ticket]: https://dsdmoj.atlassian.net/browse/AP-148
[discussion]: https://dsdmoj.atlassian.net/wiki/spaces/ATPPB/pages/1106608346/2018-09-17+Securing+Citizen+access+to+the+service

[Link to story](https://dsdmoj.atlassian.net/browse/AP-3867)