Refactor Settings
Before, new methods needed to be defined each time a new Setting was added to
the application.

This leans on `method_missing` to delegate to the Setting record.

Adding a new Setting now only involves adding the attribute to
`Setting::ATTRIBUTES`, defining labels and hint text, and running a migration to
add the new field to the Settings table.