AP-3943: Fix for citizens locked out of applications
When a citizen reaches the CYA page, the state of the application changes.
If they close the browser and return later, the application tries to run
the same starting code that involves resetting the state.  The state machine
does not allow this and, even if it did, would mean they update their bank
statements twice.

This checks for the change of state and routes them back to the CYA page

A ticket to investigate a proper save and return function for citizens
will investigate a better long term solution 🤞