Add extra delegated functions data to merits report
Before, the merits report told caseworkers the date that delegated functions
were used (if any) for each proceeding in a given application.

However this doesn't give the caseworker enough context about whether or not
date limiting the application (or any other action) would be appropriate.

This adds two new fields to the merits report for each proceeding:

1. Date delegated functions were reported on
2. Days to report that delegated functions were used

Now, caseworkers should be presented with more granualar data about delegated
functions for each proceeding —which should help them make better informed
decisions.

This change also introduces our first domain specific, non-GOVUK ViewComponent.

The component encapsulates the delegated functions information for a given
proceeding, abstracting away complexity from the main view. Unit tests help add
confidence that the information is presented correctly to the user.

One thing to note is that currently `i18n-tasks` does not support `sidecar`
ViewComponent translations. This means we can either:

1. Use sidecar translations and exclude components from `i18n-tasks`
2. Don't use sidecar translations and configure `i18n-tasks` to check components

Both options are reasonable approaches, however this change adopts the latter in
an attempt to adhere to the principle of least surprise (in the context of
Developers of this application).

[Link to story](https://dsdmoj.atlassian.net/browse/AP-3910)

Co-authored-by: Adam Goldstone <adam.goldstone@digital.justice.gov.uk>