Confirm prohibited steps is not for change of name
This adds a confirmation checkbox to the prohibited steps step of the proceeding
task list, similarly to the specific issue question.

This should act as a speedbump for providers when submitting an application, to
ensure that they do not try to submit an application for a prohibited steps
proceeding when they are trying to submit a change of name application.

[Link to story](https://dsdmoj.atlassian.net/browse/AP-3911)