AP-3802: Update shared_addresses controller to copy address
When the provider selects `yes` to the shared address question it will copy
the address from the applicant into the partner fields.  This uses a new,
`save_run_continue_or_draft` method that runs a method on the form_for target
model after validating and saving, but before redirecting. it will raise a
NoMethodError if the run_method does not exist