Update: CFE Comparison
We saw a disparity in the return values for result_summary/capital/assessed_capital
After analysing the return values it was determined that it does not affect the
Apply service.  The value is applied correctly in the calculations and we do not
explicitly display that field.  After extending the scenarios it was assessed that
the correct value is shown when a pensioner_capital_disregard is generated but
when it was not applicable Legacy returned a minus figure of the full amount and
civil returns 0 (the amount applied).  Therefore we are ignoring this field for
the comparisons going forward