Refactor the ignore_task? method in MeritsTaskService
A change to I18n was making the error message capitalised,
`Translation missing` instead of `translation missing`.
This uses a case insensitive regex to find the text at the
start of the line, regardless of capitalised status (in case
they revert the change soon!)

It also changes the handling internally, previously it
added the single error into an array and asked if "any" of
them were true!  That was probably a hold over from checking
all at once but seemed a bit pointless.  It now checks the
value and returns a single boolean response