CFE-Civil handling
There are two level of help options in CFE-Civil, Apply should
always be certificated and the documentation says that it is
the default.

Just in case that ever changes, this ensures we send the correct
value each time.

Definition from DFab:
Controlled = Provider does not apply for Legal Aid - they always
self grant.
Civil Apply's scope is only for certificated applications