Fix: Benefit breakdown view
This was intended to roll out as a single hit, but
some providers pause before submitting.  We have seen
,so far, at least one application that has a state benefit
without a description.

This fix will allow that application to complete, generate
a means report and be submitted to CCMS.  it should allow
applications in a similar state to preceed as well