Remove use of LOCAL_CLAMAV
The bin/install_clamav_on_mac was altered to install and
configure local claim to use TCP instead of Local[unix]Socket
This means local clamav server and client should work in the same way
as the production clamav server client - namely itwill use the
`clamd.container.conf` settings for TCP.

Therefore we do not need, and should no longer use, the LOCAL_CLAMAV var
and its assoicated command which uses `--fdpass` (for improved speed?!).

```
--fdpass
Pass the file descriptor permissions to clamd. This is useful if clamd is
running as a different user as it is faster than streaming the file to clamd.
Only available if connected to clamd via local(unix) socket.
```