Create load_ccms_attribute_array method
This will be used by tests to load CSV files and create
the arrays needed for large scale CCMS testing.  Rubocop
is becoming unhappy with having large in-line arrays being
created in test methods.  This will allow us to extract
them into CSV files and import them on the fly