AP-4293: Add command to build and push to new ECR repo using IRSA
The old ECR repo does not follow standard naming as other Apply
service repo. Combined with the new to implement ECR push with
IRSA as well as use an ECR lifecycle policy it was decided
to create a new repo with all three changes in place and
shift over the using that.

The old ECR repo and ENV vars can be deleted after we have been
using the new successfully for a few days.