Adapt: Update proceeding flow
When users use the back button to move back past the
proceedings search page they get stuck having to
search again.  If they have already added the single
proceedng they need this is very counter intuitive.

This PR matches the newer add another implementations
and skips to the list and add another page

It has been checked with the design team and they
are happy with the change to the flow