AP-3424: Add rake task to copy opponents to individuals
This task creates and links a new individual to each opponent
if the opponent object does not already an individual. If they have
an individual but they do not exist it creates one and links to it (so
you can delete all individuals and rerun and links will be reset). If
the opponent already has an individual it will not create a new one.

question: if opponent name changes between runs this will not copy the
changed details. Dow we need to concern ourselve with this edge case?