Fix: DB Export clear production flag on restore
The suggested script shown to the user restores the file
to the local DB but the internal metadata leaves the DB
flagged as production.  This can cause issues if the user
wants to reset or rebuild their local DB.

This PR adds the `&& rails db:environment:set RAILS_ENV=development`
command to the psql restore line.  This ensures rails is
set to be in development mode when the database is
successfully restored