AP-4284: continue filling opponent first_name and last_name
This is purely to continue to fill data on the opponent in case
we need to rollback the code base.

This commit can be removed entirely in one we are satisifed with
the polymorphic behaviour of opponent individuals or remove the logic
in the same PR that removes the first and last name attributes from
the opponent model.