AP-4350: Remove hook that retrospectively updates opponent unneeded attributes
This hook was only need to update the old attributes in case we
needed to back out the change that switched to using individuals
table to store first_name and last_name.

The `ignored_coluns` call now breaks any reference to the old
attributes, as if they had been dropped.