AP-4303: Add organisation opponent requestor
Service class to be used to search for organisation
opponents by name and/or type. This is a proof of concept
for querying the CCMS GetCommonOrg microservice endpoint.

Note: Unlike other queries this seems to require the `refdatabim`
rather than `refdatabio` xml namespace reference for sending the
search criteria attributes.