AP-4399: Prevent html tag output to screen
Negative lookahead conjuction AND negative lookbehind conjuction
(non-capturing) required to exclude any previous search term highlighting
which can result in partial HTML being output to the screen.

However, this means only the first matched term will be highlighted in certain instances
and this puts it out of sync with the actual fuzzy search matching which works and matches
on a second term even if it is only 1 character.
example: try searching for "ri prison" and only "ri" will be highlighted, despite "prison"
being the type.