Fix: Overnight purge job
The overnight purge job had been failing because the LegalFrameworkMeritsTaskList
was not being cascade deleted.  This PR updates the foreign key link
between the two models and has been tested with an anonymised DB backup.

Before the fix, the job failed with the same error seen in production. After
the fix it successfully completed