AP-4399: Reuse single cassette for org search feature
Similar to the single cassette used in specs this adds
and uses a single cassette for the organisation search feature
tests. The pattern introduced can be extended to other applicable
features where the response is expected to be the same across all
scenarios and where it is not expected to change if called more
than once within a scenario.