AP-4400: Add `exists_in_ccms` attribute to opponent
We need different functionality for organisations that
exist in CCMS already when it comes to injection (at least).

We cannot rely on the existence of the ccms_opponent_id to reliable
tell us whether the organisation already exists in CCMS as this field
is populated with a value once injection commences. It is clearer
to store this as a separate value that defaults to false and only
gets set to true if the opponent id comes from a preexisting list
of ids that exist in CCMS for that that opponent - this is specifically
the case for organisations that are stored in CCMS and have been copied
to LFA for Apply's use.