AP-4455: Adopt GDS color pallete and share via placeholder
As well as using the GDS color pallette including a tinting/lighening
of the yellow color for the background of the highlight. This is
because there are no GDS color pallete colors that suit and it
receommends using tints of 25 or 50%.

Move shared styles to placeholder and share with highlight class that
is still used by proceeding types search.