Fix: vehicle_detail error handling
This seems to have been present for a while but only triggered
today. I saw it on UAT and then shortly after on prod.

It occurs when a provider says `yes, payments remain` but then
enter text, e.g. `n/a` into the field.  One of the handlers
used `!= 0` and the other used `.zero?`.

`.zero?` fails when the value is a string but `.eql?(0)` works as
expected

This PR updates bot checks to use the same, working, style of check

`== 0` was tried but failed an erblint check suggesting `use .zero?`