AP-4452: Remove flag from task list helper and refactor
opponent_name merits task list option now goes to
the opponent_type controller. There is no new action for
the opponent_type so this logic has been removed.

However further refactorinig is possible to rename
the merits task list "name" itself and make this more
transparent. Currently we are "remapping" the opponent_name
task to opponent_type to work around this but ideally we would
just use the name opponent_type to start with.