Rubocop: address RSpec/RepeatedDescription exclusions
Remove the todo entry for RSpec/RepeatedDescription and update
the repeated descriptions in the tests.  Most of the times this
invloved clarifying the text to be more specific.  In one instance
I deleted a completely duplicated test and in another I merged the
expectations as the initial lines were identical but the check for
the error message had a slight change