AP-4505: Match redis chart hostname logic
Use a different redis host name depending on whether the
release name contains the name of the chart, "redis".

TODO: if cloud-platform-tools image is update to helm 3.7+, currently 3.6+
we could use a `Subcharts` references in a helper as below instead.

```yaml
{{/*
Function to return the name for a UAT redis chart master node host
*/}}
{{- define "apply-for-legal-aid.redis-uat-host" -}}
  {{- $redis_fullName := (include "common.names.fullname" .Subcharts.redis) -}}
  {{- $redis_cluster_domain := "cluster.local" -}}
  {{- printf "%s-master.%s.svc.%s" $redis_fullName .Release.Namespace $redis_cluster_domain -}}
{{- end -}}
```