Update i18n handling in tests
A change in rails 7.1 now raises errors when _any_ i18n lookup
fails.  Previously these tests failed because a missing translation
often meant that a negative test would pass, I expect this text
to be missing, paired with a missing translation was a definite fail!

Now that the error is raised, I have replaced the missong, often
moved, translations with the text it _should_ have been confirming
as missing