AP-4575-ext: Add cleanup to copy case confirmation form
When No is selected on the confirmation page we should
nullify the copy_case boolean flag to indicate this as
false/unset and remove the id that was search for. this
leaves the application in a state equivalent to not
having copied anything. Flows can then allow the user
to restart the process.