Rubocop: Remove StubbedMock exemptions
This PR removes uses of expect...to_return in test blocks

It, generally, replaces them with allow...to_return as this is
the correct function.  Where we were _expecting_ the return value
to be `X` I have updated the tests to reflect and set proper
expectations on the return value, or... deleted the rogue expectations!