AP-4617: Fix online dependence for bank holiday API
When wifi is off tests dependent on bank-holiday api
will fail.

This is caused by some tests being reliant on bank holiday
dates that are not seeded. A worker is used to update the dataset
so doing an inline sidekiq call and vcr stubs is also posible
but causes other workers to need stubbing too.

This change just ensure BankHoliday are always stubbed as a
single result, in the past and therefore emulates not having
any bank holidays.