AP-4694: Update SAML sessions controller
Check which page we should direct users to after successfully
signing in.  If they have already seen the start page or are
using mock saml, they should see the office page.  Otherwise
they should be directed to the start page which will then direct
them to the office sign in page