Fix: reduce overnight sentry alerts
The scheduled mail job runs at 11 and 41 minutes past the hour,
every hour.  This meant that on staging, where we shut the
database down overnight, it would error 16 times a night.

This sets the job to run 8pm to 7pm on _all_ environemnts as
that is when the service is mainly active