Fix: Fix flickering address test
The default factory for address populates address_line_one
with a building number.  Sometimes the randomly generated
building number is included elsewhere in the HTML rendering
this test failing.  This replaces the one of flickery test
with a building_name to prevent it recurring.  Changing the factory
could have affected many more tests and caused more fragility