AP-4990: Bug: Hide change links for proceedings when read_only
The check provider answers view is re-used when the provider
is waiting for the client to provide truelayer details.

A read_only flag is set and all other sections of the Check
Providers Answers page check for the presenc of the read_only
flag and hide the change links.

The proceedings and proceeding details did not, this PR adss
the check to prevent providers changing something and breaking
the flow and subsequent page views.