AP-5048: add notification banner to check merits answer page
Add notification banner to check merits answer page to show user how the information was
populated if they have copied an application. Added notification banner as an attribute
to the page template so that it is in right place within the HTML, so that the banner
is the correct width, and as we are increasingly using banners in the service.