Chore: Update docker scan action
The scan job was running at the same time as the base-docker-image
job so, if a new build is scheduled that fixes a security incident
the fixes won;t be picked up until the following day.

It now runs 30 minutes after the base-docker-image job so that it
should be testing the neweset build each day

This also adds a workflow dispatch trigger so we can run manually
if needed, like today!