AP-4840: Broken CYA with Home Address
When we changed the home address flow, we can now go straight
to the home address controller.  This had a no_state_change_required?
override that always by-passed the state change.  This meant
that when we reached the CYA page the applictaion state was
in the initialised state and could not transition to the
next state.

Reverting the home controller to match the correspondence
controller allows either one to change the state and enable
access to the CYA page.

As it is the same I have moved it into the AddressSelectionsBaseController