AP-5107: Add target_application to linked_application
The aim is to populate target application with the search result
and, if the target is already the lead for something else or is
neither a lead or an associated application, then copy it into
the lead_application field as well.

If the target application itself has a lead application, then we
copy the target.lead_application into the current.lead_application
field.  This should allow us to show the provider linking to
`target` but under the hood, setting `target.lead` as the lead
application

fixup to db udate?