AP-5297: Fix proceeding search requests for SCA
When submitting the current proceedings to LFA it expects an
array of proceeding codes, e.g. `["PB003", "PB019"]`

The previous (join) code in Apply was merging them and sending
`["PB003, PB019"]`

This caused the proceeding search code to fail and would not
match the individual codes, i.e. it could not find a proceeding
with the code "PB003, PB019" and would return non-SCA data