Retrieve firm offices for provider and compare
The existing CCMS PDA returns all offices in the providers
firm as part of the "provider offices". The new PDA API
only returns the offices of the specific individual provider (a person/user).

To understand the impact and how to handle it we want to compare all the
offices in the firm of the specific provider from the new PDA with those
coming from the old CCSM PDA.

The new PDA requires a second call to achieve this, so the relevant part
of the response body of the second call have been merged with the response
body of the first call and returned as a result object.