Update file upload form handlers
From rails 8 onwards, the model passed to form_with cannot be nil.
These submission_form variables were sometimes set to nil and this
was causing lot sof deprecation warnings.  Setting the value to
false if the variable is nil makes the form_with method happy
and does not cause errors in the tests or use