AP-5315: Fix SCA proceeding search bug
When LFA returns a code that is not present on the page, i.e.
an SCA related proceeding on an initial search, previously it
showed the failure message and returned from the loop.

This updates the proceedings javascript so that it counts
successfully displayed proceedings, using a new, `shown`, variable.
After the loop is complete it then sets the ariaText and
shows/hides the no-proceeding-items element accordingly.