AP-5336: Add housing benefit question to truelayer
Once the Disregarded Benefits section is removed, the provider
will no longer have any opportunity to declare receipt of Housing Benefit.

Therefore we will need to add the Housing Benefit question that we currently
use on the bank upload journey (asked when there are housing costs declared)
to the true layer journey for the client as well.