AP-5335: Add regular transaction component to CFE submission for truelayer
To account of the housing benefit deduction from housing payments
we need to add the housing beneing amount and frequenct to the
CFE payload. This requires including the RegualrTransactions
component which picks up the housing benefitct transaction types
values and passes it to CFE. CFE will then return the correct result.